[validator] Module functions instead of class
- Renamed validate.py to validator.py
- Validator module will simply be a collection of functions and not a
class to instantiate with a file.  There's no need to have an object.
